Re-add the wrapper type used in `v0.3` as `VolatileCell`
